app/requirements.txt
# ============================================================================
# REQUIREMENTS FOR LC EXTRACTION SYSTEM
# Python 3.10 + CUDA 11.8
# ============================================================================

# ============================================================================
# CRITICAL: Install PyTorch with CUDA 11.8 FIRST, then install this file
# 
# Run this command BEFORE pip install -r requirements.txt:
# pip install torch==2.1.2 torchvision==0.16.2 torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cu118
# ============================================================================

# Web Framework & API
fastapi>=0.104.1,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.6,<0.1.0
pydantic>=2.5.0,<3.0.0

# Core PDF and Text Processing
pdfplumber>=0.10.3,<0.11.0
PyPDF2>=3.0.1,<4.0.0
pdf2image>=1.16.3,<2.0.0
Pillow>=10.1.0,<11.0.0

# OCR Engines
pytesseract>=0.3.10,<0.4.0
easyocr>=1.7.1,<2.0.0
paddleocr>=2.7.3,<3.0.0
paddlepaddle-gpu>=2.5.2,<3.0.0

# Image Processing
numpy>=1.24.4,<2.0.0
opencv-python>=4.8.1,<5.0.0
opencv-contrib-python>=4.8.1,<5.0.0
scikit-image>=0.22.0,<0.23.0

# Document Processing
python-docx>=1.1.0,<2.0.0
docx2txt>=0.8,<0.9
openpyxl>=3.1.2,<4.0.0
python-pptx>=0.6.23,<0.7.0

# Data Processing
pandas>=2.1.3,<3.0.0
tabulate>=0.9.0,<0.10.0

# AI / NLP (Install PyTorch FIRST with the command above!)
transformers>=4.36.2,<5.0.0
accelerate>=0.25.0,<0.26.0
sentencepiece>=0.1.99,<0.2.0
tokenizers>=0.15.0,<0.16.0
safetensors>=0.4.1,<0.5.0

# Utilities
tqdm>=4.66.1,<5.0.0
python-dotenv>=1.0.0,<2.0.0
python-dateutil>=2.8.2,<3.0.0
pyyaml>=6.0.1,<7.0.0
loguru>=0.7.2,<0.8.0

# Performance & Optimization
joblib>=1.3.2,<2.0.0
cachetools>=5.3.2,<6.0.0